requirements.txt
-e .
black==23.7.0
boto3==1.35.29
celery[sqs]==5.4.0
ffmpeg-python==0.2.0
flake8==6.1.0
ipykernel==6.29.5
loguru==0.7.2
mypy==1.5.1
openai==1.55.2
pandas==2.2.2
pillow==11.0.0
pre-commit==3.7.0
pydub==0.25.1
python-dotenv==1.0.1
toml==0.10.2
yt-dlp==2024.11.18
